Migrate Mutation.enableProjectNameInGitHubCheck
